// discode-web/src/app/home/models/member.model.ts
export class Member {
    chatMemberId: BigInteger | undefined;
    chatId: BigInteger | undefined;
    userId: BigInteger | undefined;
    username: string | undefined;
    firstName: string | undefined;
    lastName: string | undefined;
    imagePath: string | undefined;
    status: string | undefined;
}